+++ apps/mobile/src/main/java/com/dpanger/vehicles/HomeActivity.kt
@@ -4,11 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import androidx.hilt.navigation.compose.hiltViewModel
-import com.dpanger.vehicles.features.manufacturers.ManufacturersScreen
 import com.dpanger.vehicles.ui.themes.VehiclesTheme
 import dagger.hilt.android.AndroidEntryPoint
