+++ features/manufacturers/src/main/java/com/dpanger/vehicles/features/manufacturers/ui/ManufacturersList.kt
@@ -1,4 +1,4 @@
-package com.dpanger.vehicles.features.manufacturers
+package com.dpanger.vehicles.features.manufacturers.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -7,6 +7,7 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import com.dpanger.vehicles.features.manufacturers.viewmodel.UiManufacturer
 import com.dpanger.vehicles.ui.themes.VehiclesTheme
 import com.dpanger.vehicles.uicomponents.components.lists.SingleLineItem
 import kotlinx.collections.immutable.ImmutableList
