+++ features/manufacturers/src/main/java/com/dpanger/vehicles/features/manufacturers/ui/ManufacturersScreen.kt
@@ -1,8 +1,10 @@
-package com.dpanger.vehicles.features.manufacturers
+package com.dpanger.vehicles.features.manufacturers.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.dpanger.vehicles.features.manufacturers.viewmodel.ManufacturersUiState
+import com.dpanger.vehicles.features.manufacturers.viewmodel.ManufacturersViewModel
 import com.dpanger.vehicles.uicomponents.components.error.ErrorMessage
 import com.dpanger.vehicles.uicomponents.components.progress.ProgressIndicator
