+++ features/manufacturers/src/main/java/com/dpanger/vehicles/features/manufacturers/viewmodel/UiManufacturer.kt
@@ -1,4 +1,4 @@
-package com.dpanger.vehicles.features.manufacturers
+package com.dpanger.vehicles.features.manufacturers.viewmodel
 
 data class UiManufacturer(
     val id: String,
